Removed the types modules and moved all code to the appropriate modules
