+++ src/config/command.rs
@@ -2,7 +2,7 @@ use crate::config::{
     definitions::VERSION, Auth, AuthValidation, BldConfig, BldLocalConfig, BldRemoteConfig,
 };
 use crate::helpers::term;
-use crate::types::BldCommand;
+use crate::cli::BldCommand;
 use clap::{App, Arg, ArgMatches, SubCommand};
 
 static CONFIG: &str = "config";
