+++ src/monit/client.rs
@@ -1,4 +1,4 @@
-use crate::types::MonitInfo;
+use crate::monit::MonitInfo;
 use actix::io::{SinkWrite, WriteHandler};
 use actix::{Actor, ActorContext, AsyncContext, Context, Handler, StreamHandler, System};
 use actix_codec::Framed;
