+++ src/monit/command.rs
@@ -3,7 +3,8 @@ use crate::helpers::errors::auth_for_server_invalid;
 use crate::helpers::request::headers;
 use crate::helpers::term::print_error;
 use crate::monit::MonitClient;
-use crate::types::{BldCommand, MonitInfo};
+use crate::cli::BldCommand;
+use crate::monit::MonitInfo;
 use anyhow::anyhow;
 use actix::{io::SinkWrite, Actor, Arbiter, StreamHandler, System};
 use awc::Client;
