+++ src/monit/mod.rs
@@ -1,5 +1,7 @@
 mod client;
 mod command;
+mod messages;
 
 pub use client::*;
 pub use command::*;
+pub use messages::*;
