+++ src/run/sync/runner.rs
@@ -1,8 +1,8 @@
 use crate::config::definitions::{GET, PUSH, VAR_TOKEN};
 use crate::config::BldConfig;
+use crate::run::CheckStopSignal;
 use crate::persist::{Execution, Logger, NullExec};
 use crate::run::{BuildStep, Container, Machine, Pipeline, RunsOn};
-use crate::types::CheckStopSignal;
 use std::collections::HashMap;
 use std::future::Future;
 use std::pin::Pin;
