+++ src/server/endpoints/auth_redirect.rs
@@ -1,5 +1,11 @@
-use crate::types::AuthRedirectInfo;
 use actix_web::{get, web, HttpResponse};
+use serde_derive::Deserialize;
+
+#[derive(Deserialize)]
+pub struct AuthRedirectInfo {
+    pub code: String,
+    pub state: String,
+}
 
 #[get("/authRedirect")]
 pub fn auth_redirect(web::Query(info): web::Query<AuthRedirectInfo>) -> HttpResponse {
