+++ src/server/endpoints/push.rs
@@ -1,6 +1,6 @@
 use crate::run::Pipeline;
 use crate::server::User;
-use crate::types::PushInfo;
+use crate::push::PushInfo;
 use actix_web::{post, web, HttpResponse, Responder};
 use std::fs::{remove_file, File};
 use std::io::Write;
