--- src/types/auth.rs
@@ -1,7 +1,0 @@
-use serde_derive::Deserialize;
-
-#[derive(Deserialize)]
-pub struct AuthRedirectInfo {
-    pub code: String,
-    pub state: String,
-}
