--- src/types/mod.rs
@@ -1,17 +1,0 @@
-mod auth;
-mod command;
-mod exec;
-mod monit;
-mod push;
-mod recv;
-
-pub use auth::*;
-pub use command::*;
-pub use exec::*;
-pub use monit::*;
-pub use push::*;
-pub use recv::*;
-
-use yaml_rust::Yaml;
-
-pub static EMPTY_YAML_VEC: Vec<Yaml> = Vec::new();
